refactor(bezier-codemod): separate enum transform function, add ZIndex.Auto to transform target
