+++ packages/bezier-codemod/src/transforms/enum-member-to-string-literal/transform.ts
@@ -1,60 +1,24 @@
+import { type SourceFile } from 'ts-morph'
+
 import {
-  type SourceFile,
-  SyntaxKind,
-} from 'ts-morph'
+  type EnumTransformMap,
+  transformEnumMemberToStringLiteral,
+} from '../../utils/enum.js'
 
-type EnumTransforms = Record<string, Record<string, string>>
-
-export function transformEnumMemberToStringLiteral(sourceFile: SourceFile, enumTransforms: EnumTransforms) {
-  const transformedEnumNames: string[] = []
-
-  sourceFile
-    .getDescendantsOfKind(SyntaxKind.PropertyAccessExpression)
-    .forEach((node) => {
-      const firstIdentifier = node.getFirstChildByKind(SyntaxKind.Identifier)
-      const lastIdentifier = node.getLastChildByKind(SyntaxKind.Identifier)
-
-      const enumName = firstIdentifier?.getText()
-      const enumValue = lastIdentifier?.getText()
-
-      if (!enumName || !enumValue) { return }
-      if (!Object.keys(enumTransforms).includes(enumName)) { return }
-
-      const newEnumMemberValue = enumTransforms[enumName]?.[enumValue]
-      const ancestor = node.getFirstAncestor()
-
-      if (!newEnumMemberValue) { return }
-
-      if (ancestor?.isKind(SyntaxKind.JsxExpression)) {
-        ancestor.replaceWithText(`'${newEnumMemberValue}'`)
-      } else {
-        node.replaceWithText(`'${newEnumMemberValue}'`)
-      }
-
-      transformedEnumNames.push(enumName)
-    })
-
-  if (transformedEnumNames.length > 0) {
-    sourceFile.fixUnusedIdentifiers()
-    return true
-  }
-
-  return undefined
+const ENUM_TRANSFORM_MAP: EnumTransformMap = {
+  ProgressBarSize: {
+    M: 'm',
+    S: 's',
+  },
+  ProgressBarVariant: {
+    Green: 'green',
+    GreenAlt: 'green-alt',
+    Monochrome: 'monochrome',
+  },
 }
 
 function enumMemberToStringLiteral(sourceFile: SourceFile): true | void {
-  const enumTransforms: EnumTransforms = {
-    ProgressBarSize: {
-      M: 'm',
-      S: 's',
-    },
-    ProgressBarVariant: {
-      Green: 'green',
-      GreenAlt: 'green-alt',
-      Monochrome: 'monochrome',
-    },
-  }
-  return transformEnumMemberToStringLiteral(sourceFile, enumTransforms)
+  return transformEnumMemberToStringLiteral(sourceFile, ENUM_TRANSFORM_MAP)
 }
 
 export default enumMemberToStringLiteral
